debugging

+++ makeCSV.py
@@ -38,7 +38,7 @@ def get_channels(channels,bfactor):
     """
     ch_array = np.frombuffer(channels,dtype=np.uint8)
     print(ch_array)
-    rebin_array = ch_array.reshape(len(ch_array)/bfactor,bfactor).sum(1)
+    rebin_array = ch_array.reshape(int(len(ch_array)/bfactor),bfactor).sum(1)
     return rebin_array
 
 def get_rounded_time(t):
